CLM-001 First Commit
Add messaging library with producer interface and Message class

This commit adds a new common library for messaging.
A Message class is created with 'MessageType' as an enum and 'id' as a String.
Also, a Producer interface is created with 'sendMessage' method.

The Maven settings for the project are established in a new pom.xml file.
The build process is updated to ignore certain temporary files and folders.